outsourced ODCreateMoveSnapping
